Fix: rework getSize so that it casts unknown types to a number
Trying to predict types and returning el if its type is not known is not
a good logic, since, for example, functions or custom objects could be
returned as is and create problems when used in comparisons.

Instead, cast all unknown types to a number representing their boolean
form and assure that getSize will always return a number.