Fix lead and trail by using size as is
Focusing on producing 2-digit numbers, the current version uses the
number of digits in the receiving number and adds 1 to it, thinking
solely on receiving 1-digit numbers.

Obviously, the receiving number could already be a 2-digit number,
making the function actually produce a number with 1 digit more than the
intended.

The fix is, of course, just using size as is and setting its default
value to 2.