Use getParDirName() in getDirAsArr()
In order for stack() to include the parent's path, it must be called in
a function that will be called by the parent, directly or indirectly.

When the __pardirname const is used, the const calls stack() in
Richierich itself, thus preventing the parent's path from being included
in its result.