Remove the useless __parfilename and __pardirname consts
By providing a callsite stack called inside Richierich itself, these
consts will never really do what they're supposed to do: provide the
filename and the directory of the parent script (i.e. the script using
Richierich).

For that, the callsite stack() method needs to be called in a function
that will be called by the parent script, directly or indirectly.